pyproject.toml
[tool.poetry]
name = "core-data-covid-project"
version = "0.1.0"
description = "Data loader part of the mid-project for the Data Science bootcamp from Core Code School"
authors = ["Ale Sanchez <alesanchezmedina@gmail.com>"]
license = "The Unlicense"

[tool.pyright]
pythonVersion="3.9"
typeCheckingMode="basic"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.0"
beautifulsoup4 = "^4.9.3"
click = "^8.0.1"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
psycopg2 = "^2.9.1"
Unidecode = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
setuptools = "^57.4.0"
isort = "^5.9.2"
pylint = "^2.9.5"
flake8 = "^3.9.2"
bandit = "^1.7.0"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
